src/main/java/com/cognizant/triton/ddd/valueobjects/Money.java
package com.cognizant.triton.ddd.valueobjects;

import java.io.Serializable;
import java.math.BigDecimal;
import java.util.Currency;
import java.util.Objects;

public class Money implements Serializable, Comparable<Money> {
    private final BigDecimal amount;
    private final Currency currency;

    public Money(BigDecimal amount, Currency currency) {
        this.amount = Objects.requireNonNull(amount);
        this.currency = Objects.requireNonNull(currency);
    }

    public Money add(Money other) {
        assertSameCurrency(other);
        return new Money(amount.add(other.amount), currency);
    }

    public Money subtract(Money other) {
        assertSameCurrency(other);
        return new Money(amount.subtract(other.amount), currency);
    }

    private void assertSameCurrency(Money other) {
        if (!other.currency.equals(this.currency))
            throw new IllegalArgumentException("Money objects must have the same currency");
    }

    @Override
    public int compareTo(Money o) {
        return 0;
    }

    public BigDecimal value(){
        return amount;
    }
}